+++ scripts/run_tests.sh
@@ -17,5 +17,5 @@ function display_result {
 pep8 .
 display_result $? 1 "Code style check"
 
-py.test
+py.test $@
 display_result $? 2 "Unit tests"